requirements.txt
quotexpy
python-dotenv
fastapi
uvicorn
